Optimize setting parameters for transactions
With the recent refactoring of the Rack middleware, I changed how we set
the request parameters on the transaction. Previously, we read the
parameters from the request object in the transaction when we sampled
it. Not always.

Now that we use the currently active transaction when a transaction is
already active, I needed to update how we set parameters using the
`set_params`/`set_params_if_nil` helpers. These helper methods require
us to pass in the parsed parameters, potentially making every request
slower because it has to fetch and parse the request parameters.

Add a block argument to the `set_params`/`set_params_if_nil` helpers to
fetch and parse the parameters only when we sample the transaction. This
change should reduce the AppSignal overhead for non-sampled
transactions.

I have no specific preference if the argument or block is leading. I
chose the argument to be leading.