Add set_custom_data instrumentation helper (#1149)
I saw we tell people in our docs to use the
`Appsignal::Transaction#set_sample_data` method. Let's not do that.

Here's a helper to hide all those internals away and make it less likely
to break when setting custom data as sample data.

I considered adding logic to merge the custom data, but merging these
values is quite tricky because it can be both a Hash and an Array as the
root object. It's not something I want to think about right now. If we
want to add this in the future, we can also name that helper
`add_custom_data` to differentiate between setting and merging the
custom data.

Closes #1147