Add Appsignal.set_headers helper method
I want to deprecate and remove the `Transaction#set_session_data`
method. To be able to remove it without breaking things for applications
calling it directly, add helpers for all types of sample data. This
commit adds request headers helpers. This is similar to our Node.js
sample data helpers.