Refactor Shoryuken to not use monitor_transaction
Refactor the Shoryuken integration not to use the `monitor_transaction`
helper. A helper we want to deprecate.

The implementation is mostly the same, except how we set the attributes.
I have changed it to store them as tags. Previously we set the
attributes as "sample data metadata". This is another way to set
metadata/tags in the transaction API. All three data types end up in the
"overview" tag box, so it doesn't really matter how we set the values.

The only thing is that the sample metadata seems to support more value
types than tags. I changed this in PR #1169 by accepting boolean values
in tags.

For reference, here's a single and matched message's attributes
Shoryuken can receive.

```ruby
# Single
{"queue"=>"default",
 "SenderId"=>"AIDAIT2UOQQY3AUEKVGXU",
 "SentTimestamp"=>"1720542601233",
 "ApproximateReceiveCount"=>"1",
 "ApproximateFirstReceiveTimestamp"=>"1720542601319",
 "message_id"=>"9b40dbd8-5c26-40ea-8ded-6d68428072fb"}

# Batched
{"queue"=>"batched",
 "SenderId"=>"AIDAIT2UOQQY3AUEKVGXU",
 "SentTimestamp"=>"1720542649820",
 "ApproximateReceiveCount"=>"1",
 "ApproximateFirstReceiveTimestamp"=>"1720542649833",
 "batch"=>true}
```