Do not use monitor_transaction for Delayed::Job
Refactor the Delayed::Job integration not to use the
`monitor_transaction` helper. A helper we want to deprecate.

The implementation is mostly the same, except how we set the attributes.
I have changed it to store them as tags. Previously we set the
attributes as "sample data metadata". This is another way to set
metadata/tags in the transaction API. All three data types end up in the
"overview" tag box, so it doesn't really matter how we set the values.

[skip changeset] because it's an internal refactor.