Deprecate `set_http_or_background_queue_start`
Deprecate the `Transaction#set_http_or_background_queue_start` method so
that the transaction doesn't need to know about any 'request' set on the
transaction. The instrumentations themselves can set the queue start
using the `Transaction#set_queue_start` method.

Move the logic for conversion of the queue start time from request
headers to a utils class so it can be accessed by the instrumentations.