Set Action Cable request metadata using helpers
Update the Action Cable instrumentation to not rely on the request
object but set the request metadata using the `set_*` helpers.

This uncovered a wrong assertion in the test suite when an active
transaction was already set for the subscribe and unsubscribe callbacks.
This setup made the spec not use the request object created in the
instrumentation but in the spec setup.

As a result, looks like these scenarios (for Rails' test suite) doesn't
actually report any request params.

[skip changeset]
[skip review]