Set Action Cable request ID as a tag
Do not set the request ID as the transaction ID. Let's use generated
transaction IDs for that. If the "magic" tag `request_id` is set, it has
the same behavior as the legacy behavior of using the transaction ID for
this. This implementation is more explicit and doesn't cause conflicting
transaction IDs.

[skip changeset]
[skip review]