Set Sidekiq job ID as a tag
Do not set the job ID as the transaction ID. Let's use generated
transaction IDs for that. If the "magic" tag `request_id` is set, it has
the same behavior as the legacy behavior of using the transaction ID for
this. This implementation is more explicit.

The naming of the `request_id` tag doesn't fit well in the context, but
that's the only tag our frontend listens too right now to connect to our
logging feature.

[skip changeset]
[skip review]