Remove the warn_once_then_debug from the logger (#1185)
This method was originally added to stop logging many messages in Grape
apps nested in Rails apps. That instrumentation didn't work well in
older versions of the Ruby gem. That's been fixed with the recent Rack
middleware refactors.

Let's not treat this warning as an annoyance anymore and instead treat
it as a symptom of an issue we should fix. It should be happening
anymore in nested apps and is probably an indication of incorrectly
setup instrumentation.

See issue #584

[skip changeset] because it's an internal refactor