Remove Action Cable dependency on request object
Don't pass the request object to the transaction, but set the request
data via the instrumentation helpers. This follows the instrumentation
pattern we've added in other integrations.

Add more assertions to tests to make sure it does what we expect.
There's no queue time for these transaction as it sent from a websocket,
so no queue time.