Make action argument required for monitor helper
Avoid potential issues with the action name not being set and no
transaction being reported.

For most scenarios it will be fine to set the action name before the
block is instrumented. For the more advanced scenarios the action name
argument can be set to `nil` or `:set_later` and call
`Appsignal.set_action` in the block to set the action name.