Add dedicated clear_config! helper
Instead of writing `nil` to the `Appsignal.config` accessor, add a
helper for this. This makes it more explicit what's happening when
it's set to `nil`.

I also want to remove the `Appsignal.config=` writer and this provides
another way to unset the config.

It's private because no one other than us should call it.