Add Appsignal.configure helper
Create a new interface to configure AppSignal.

The `Appsignal.config = Appsignal::Config.new` method has some problems
with it.

- Apps can overwrite the config at any time, which causes problems like
  AppSignal boot loops.
- The environment and root path need to be specified all the time,
  overwriting any previously set value.
- The config given to the helper is overwritten by the file config and
  environment variables config. Which goes against what we decided in:
  https://github.com/appsignal/integration-guide/issues/73

This new `configure` helper will allow apps to merge config with the
already loaded config from the file config and environment. The config
set with the `configure` helper is leading, as it can access the already
loaded file and environment config, and modify it.

I had to change how to config is loaded. I moved it to a new
`load_config` method that's called if apps use the old
`Appsignal.config = Appsignal::Config.new(...)` way of configuring
AppSignal. If apps use the new `Appsignal.configure { ... }` way, this
is called by that helper method at the right time.