Fix duplicate response body instrumentation events
Multiple response body instrumentation events could show up if between
two AppSignal instrumentation middlewares there is another middleware
that wraps around the response body too. That way we couldn't detect if
the response body was of a certain type anymore, leading to duplicate
response body instrumentation events.

To fix this, add another flag to the Rack request env to flag that the
response is already instrumented and we don't need to apply the
instrumentation again.