Allow merging of sample data
We've received requests in the past to merge sample data. Our API was
inconsistent in merging some sample data like tags, but not others
(params, headers, session data and custom data).

Support merging sample data for all sample data. All helpers now are
called `add_<data type>`. I've added aliases for the `set_<data type>`
methods.

This merging behavior is the new default, which is a breaking change.

I've added a SampleData helper class to handle the sample data being
merged. Array and Hashes are supported, same as in our extension. Other
types are not supported.

If the root value type switches between a Hash and Array between values
being merged, the new value is leading, discarding the previous value.

I've kept the behavior that if an application sets custom
params/headers/session data/custom data, we do not set it from our
integrations. This would merge it with the custom set data from the app,
which I don't know if we want to do that.