Alias `set_error` to `add_error` method
Instead of having `add_error` always add the errors to the additional
errors array, and then "promoting" the last of those errors to an error
in the current transaction later, when the transaction is completed,
do it when an error is first added using `add_error`.

To avoid calls to `set_error` overriding this error instead of adding
a new one, make `set_error` into an alias to `add_error`. Move the
logic to set an error in the transaction into a private `_set_error`
method.