Fix loader defaults from overriding other sources
The loader defaults would override the `APPSIGNAL_APP_ENV` env var,
making it no longer leading. It would also override the initial env
given to the Config initializer, which it shouldn't do.

I made the loader defaults logic in the Config class too complex. I
wanted to put all the logic in one place, but it created this mess.

This change reduces the complexity a bunch by not supporting any default
config options from the loaders except `env` and `root_path` for now.

Instead of determining the env and root path from all the different
sources in the Config initializer, move them to the place the config is
initialized (and move them to a helper).

After this change the loader defaults aren't considered anymore if an
app calls `Appsignal::Config.new` manually, but that's deprecated and
didn't work as expected. Since `Appsignal::Config.new` requires the
`root_path` and `env` to always be given to the initializer, it wouldn't
do anything anyway.

Fixes #1213