Ignore Puma::ConnectionError when reading body
We got another report of the `Puma::ConnectionError` error being
reported when upgrading our Ruby gem.

It is reported by our response body wrapper that's called when
instrumenting response bodies.

In this case the error was:

Puma::ConnectionError: Socket timeout writing data

The cause of this error was this error:

> Errno::EPIPE: Broken pipe

This doesn't look like an error people can act on to fix it, let's
ignore it instead.

Related issue https://github.com/appsignal/support/issues/315