Keep the first ten errors
Before this change, the last ten additional errors would be kept,
alongside with the first (non-additional) error. This was a strange
mix of behaviours that would have been hard to document. Instead,
we can now document that we keep the first ten errors.

The language emitted in the warning has been modified to avoid any
mention of "additional" errors.