Make global helpers work in error method callbacks
When using `Appsignal.report_error` and `Appsignal.send_error` (and
`Appsignal.set_error`, more on that later), it was required to customize
the metadata by calling the `set_*` methods directly on the Transaction
object.

This is the only time that is needed in our public API and wasn't ever
really documented what methods applications can call, and how they might
be different from the helpers on `Appsignal`.

(I don't think there are any differences now, but there were in the past
when we didn't have global helpers for custom data, parameters, headers
and session data.)

To make matters simpler for our end-user, make the `Appsignal.set_*`
helpers work in the callback of `Appsignal.report_error` and
`Appsignal.send_error` by temporarily switching the current transaction
for the duration of that callback block.

In the `Appsignal.set_error` helper callback, applications could already
use `Appsignal.set_*` helpers, because it can only set errors an active
transaction so I only updated the API docs for it.