Fix app env being reported as empty brackets
In PR #1214, I added the helper `Appsignal::Config.determine_env`. When
no valid env could be found it would return `[]`.

If the application is configured through env vars, except the app env,
it would report the env as `[]`. That's because it would return the list
of loader defaults if no env was found. (It could also return a list of
loaded loaders.)

Fix this fallback by returning `nil` if no env is found.