Allow `add_error` to receive a block
Before this change, helpers like `send_error` and `report_error`
took a block as an argument, which they applied immediately in the
error transaction or in the current transaction.

After this change, this behaviour is part of `add_error` itself, with
`send_error` and `report_error` using it under the hood. When a block
is given as an argument when adding an error, the block is stored
alongside the error, and applied only to the transaction that reports
that specific error.

Change the behaviour of `report_error` so that it does not force
`send_error`-like behaviour when a block is given. It will only behave
like `send_error` if there is no current transaction.