Fix Padrino require error
When Padrino is loaded it will raise an error about `Rack::Server` not
being found. That's because the constant has been renamed to
`Rackup::Server` in newer versions of Rack and Rackup.

The latest version of Padrino is not updated for this yet, so catch the
error if it occurs. This makes our detection a bit less accurate, but at
least it won't crash the installer.