Install gem when no known framework found
When no supported framework is installed, the installer would exit. It
would try to send the demo data, but this would always fail because
there is no valid AppSignal config chosen.

Change the installer that it will configure AppSignal when no supported
framework is found, and link to the docs for a list of other libraries
we integrate with. Some manual setup is required but we don't known
which library docs to link to.

Update the installer output to always point people back to installation
instructions in the browser.

Part of #315