Add `at_exit` error reporter (#1218)
We can listen to Ruby's `at_exit` to do things right before the process
exits.

In this hook, see if any error was unhandled and will crash the process
by checking the `$!` variable. If this is the case and the error was not
previously reported on a transaction, report it.

Add an `AppSignal.stop` to make sure the transaction gets flushed before
the process exits. We should also add some docs to tell people to add a
sleep to processes that are the only process on a short-lived host like
a container so it shuts down immediately after the script ends.

For example:

```ruby
at_exit { sleep 5 }
```