Add option to customize config root path
I added an option before to customize the root path via the `app_path=`
writer on the ConfigDSL. However, this is set too late, after the file
config is already read, so it has no effect.

Add a keyword argument to the `Appsignal.configure` helper so that it
can still be customized for scenarios where the current working
directory is not the place to look for the config file.