Deprecate Appsignal.listen_for_error helper (#1229)
As discussed, remove the `listen_for_error` helper. It doesn't allow for
as much customization as `report_error` does. We're not going to add a
ton of arguments to `listen_for_error` to allow all kinds of metadata to
be set that way. Instead, let's remove it in the next version.

https://appsignal.slack.com/archives/C07BBHJCNP9/p1722354302496879