Merge loader config options with default config
Revert the changes in #1214 and make them better.

I want to make the Rails railtie use loader defaults as well, so that it
can set config defaults that are also used by any `Appsignal.configure`
calls from Rails initializers. Currently we only set the defaults if
there is no `Appsignal.configure` call in the Rails initializers.

I have separated the behavior for env, root_path and the config options
even more so that the env and root_path can't be overwritten by the
loader defaults. The loader env and root_path defaults only read in the
`Config.determine_env` and `Config.determine_root_path` helpers.