Load Rails config in diagnose CLI
Make the diagnose CLI more useful for Rails apps by loading the Rails
app to fetch the configuration.

This fixes a couple things:

- Include config defined in Rails initializers.
- Accurately include Rails default config set by the Railtie. Previously
  we duplicated the behavior in the diagnose CLI, but that could diverge
  when changed.

To make this work, require the `config/environment.rb` file which
initializes Rails, loads the initializers and does most things but start
the Rails server.

The diagnose CLI will also no longer initialize the AppSignal config
with the `Appsignal._config =` method, which I'm trying to remove
entirely.

Do not start AppSignal in `Appsignal.start` if the `_APPSIGNAL_DIAGNOSE`
env is set, so that loading the Rails app in the diagnose CLI doesn't
start AppSignal (the agent specifically).

There's a bunch of changes for the test suite so there's only one Rails
app loaded, because Rails only allows one at a time, and it can only be
initialized once.

Fixes #510