Refactor diagnose CLI specs (#1233)
Don't use a Config object in the spec to determine the env, root path
and app name. With the auto detection not all present in the config
class itself, it's a bit annoying to deal with. I refactored it to a
plain Hash with the important bits so it's easier to configure the HTTP
requests in different scenarios.