Don't auto run diagnose in Rails app context
When running the diagnose spec in the Rails gemfile contexts, don't let
it auto detect the working dir as a Rails app. That requires us to keep
a bunch of test wide different config defaults in mind.

This commit removes the Rails context from the test suite, and in the
next commit I'll make separate tests for Rails.