Include loader and DSL source in diagnose report (#1235)
Show all sources for config in the diagnose report. I didn't add this in
the PRs that added the loaders and DSL config, because no part of the
diagnose command would load code that can modify these sources. Now I'm
working on loading a Rails app in the diagnose command, so it's good to
include these sources for when we also want to include the config from
the Rails app.