Use report_error in Rails error subscriber
The Rails error subscriber had already some kind of duplicate
transaction behavior, when we didn't support multiple errors.

Remove the transaction duplication logic and use
`Appsignal.report_error` instead, to report the error. This will
duplicate the transaction metadata more accurately when the transaction
is closed, rather than whenever the error is reported. It will also
duplicate all the sample data, not just tags.

If a "context" is given to `Rails.error.handle`, it will overwrite the
data on the transaction with the given data. This is mostly useful for
errors outside of an instrumented context, when a new transaction is
created by `Appsignal.report_error`, and shouldn't be necessary for most
other cases.

[skip changeset]