Transmit check-in events in thread
Implement a check-in event scheduler, which schedules check-in events
to be transmitted in a separate thread, with a minimum wait period
of ten seconds between requests, and a wait period of a tenth of a
second before the first request, referred to as "debounce" periods.

This is a relatively minor improvement for the existing cron
check-ins, but it is a requirement for the heartbeat check-ins, both
to avoid slowing down customers' applications with blocking requests,
and to avoid misuse of the feature from spamming our servers.

The scheduler also acts as a deduplicator, removing "similar enough"
check-in events -- again, not particularly interesting for cron
check-ins, but a requirement to minimise damage when heartbeat
check-ins are misused.

Implement support for NDJSON payloads in the transmitter in order to
send all the check-ins in a single request.

The whole setup is lazily instantiated, so no threads will be spawned
until a check-in event actually needs to be transmitted. Internally,
a queue is used to communicate with the thread, which is suspended
until a message is pushed into the queue.

A second thread is used to keep track of the debounce period, and is
killed after a debounce period does not cause any events to be sent.

When `Appsignal.stop` is called, the scheduler is gracefully stopped,
blocking the call to `Appsignal.stop` until it has transmitted all
enqueued check-in events. During this time, new check-in events
cannot be scheduled.