Add helper to unset parameters sample data
Our own projects unset the parameters to a nil or empty value sometimes
for specific endpoints.

Add a helper to replace this old method of doing it:

```ruby
Appsignal::Transaction.current.params = {}
```

I've kept the method private for now. We can make it stable if we're
happy with how it works and the naming of it, which I didn't spend too
much time on. The whole `.value?` and `.empty?` being able to contradict
each other is a bit weird.