Ignore closed pipe error in response body wrapper
Do not report closed pipe errors when instrumenting the response body in
the body wrapper classes. We've gotten reports of this error being
reported when a user clicks a link, but before the page managed to load,
click away to another page. We shouldn't report this error when this
happens.