Fix error calling body methods on BodyWrapper
When a method exists on the wrapped body but not on
`Appsignal::Rack::BodyWrapper` forward the method call to the body, like
the `Rack::BodyProxy` does as well.

This way, we mimic the behavior better, and applications won't know
they're talking to our wrapper. They shouldn't have to change their code
to accommodate the AppSignal gem.

The risk is that an application calls the wrong body method (`each`,
`call`, `to_ary` or `to_path`), and the body gets processed by bypassing
our wrapper. This risk is quite small. Our wrapper selects the best
wrapper depending on the methods the original body supports. In the
scenario a body supports both `each` and `call`, and we wrap it with the
`each` wrapper, but if the application calls `call`, we will not
instrument the body.

Fixes #1265