Extract Rack request metadata helpers to class
Create a new class for the Rack request metadata extraction so it can be
reused in other places more easily too.

Some of the specs are duplicated. I've moved all the specs for edge
cases to the class with the actual implementation. The
AbstractMiddleware just tests the basics that we expect that class to
handle.