Report low-level Puma errors
When an error is reported by Puma's `lowlevel_error` reporter, it
doesn't arrive at our `EventHandler`'s `on_error` callback or through
the `AbstractMiddleware` instrumentation. This resulted in the report in
issue #1269 of an error not being reported.

Report the error through a patch on the `Puma::Server` class. I first
wanted to add it as described in their docs:
https://puma.io/puma/index.html#error-handling
This method only works if you can configure it through the Puma config
file. It would also require us to determine the response in the block
given to that setting. This patch will call the super method, which will
determine the response and also allows applications to specify their own
`lowlevel_error` handler.

Use the `Appsignal::Rack::ApplyRackRequest` helper to set the request
metadata on the transaction, so it's hopefully a bit easier for people
who view errors reported this way to debug the issue. This error doesn't
flow normally through our instrumentation, so it needs this metadata set
manually like this. The transaction is closed through the
`RACK_AFTER_REPLY` hook set in the EventHandler. Usuallly, the
EventHandler still creates the transaction for the request. If the
EventHandler doesn't create the transaction because it's not set up
correctly, the `report_error` helper will create a new transaction for
this error.

Since version 4 of the Ruby gem we've gotten a couple of reports of
applications suddenly reporting new errors, usually internal errors
developers can't really fix. This patch will ignore a couple of internal
errors from Puma as well to prevent those issues from appearing in the
next couple of days.

Closes #1247
Fixes #1269