Format Date and Time objects stored in sample data
Format the Date and Time objects in a human-friendly way. Previously,
dates and times stored in sample data, like session data, would be shown
as `#<Date>` and `#<Time>`.

They will now be show as `#<Date: 2024-09-11>` and
`#<Time: Time: 2024-09-12T13:14:15+02:00>`.
(The time zone will be based on the setting in Ruby.)

The UTC offset is included in the Time object on purpose. Casting it to
UTC will lose the timezone information.