Fix ThreadError on calling stop in Signal.trap
When `Appsignal.stop` is called inside a `Signal.trap` block, it raises
a `ThreadError`. This is because our `CheckIn::Scheduler` calls
`Mutex#synchronize` and this is not allowed in Ruby.

> Unsafe methods to call in Signal.trap blocks > Mutex#lock,
Mutex#synchronize and any code using them are explicitly > unsafe. This
includes Monitor in the standard library which uses Mutex > to provide
reentrancy.

Source: https://docs.ruby-lang.org/en/master/signals_rdoc.html

To fix this, wrap the entire `Appsignal.stop` method contents in a
Thread, and join it at the end. This does exactly the same, but in a
thread, and makes Ruby happy.

This solution is inspired by this StackOverflow answer:
https://stackoverflow.com/a/28052251/543643

To test this, I didn't want to declare `Signal.trap` inside the RSpec
process for a unit test. Instead created an integration test that runs a
script that has a signal trap and calls `Appsignal.stop` from within
that signal trap.
The Runner helper class implementation is largely copied from the
Diagnose integration test submodule.