Remove backtrace cleaner filter after test (#1303)
I ran into an issue adding a spec with the line number "123". It would
be replaced with "1?3" whenever Rails was loaded. That's because this
spec left the filters in the cleaner's config.

Disable the filter after the spec has completed. I could go into the
backtrace cleaner internals and remove the lambda from the `@filters`
instance variable, but this solution with our global store that's only
set to true for this spec is fine. The store is cleared after every
spec.

[skip changeset]