Assert the logger works in Capistrano
Follow up on the previous commits to remove the logger instance from the
Config class, which was set by Capistrano.

It didn't actually log anything before, as it wrote to a `StringIO.new`.
Log now to the configured logger based on the application's config.

I see no harm in making the logger work. By default it will log to the
`appsignal.log` file. If apps have configured it to log to stdout, they
have deliberately configured that so there should be no problem.