Fix Pathname to String conversion error (#1310)
In Rails apps the app root path is a Pathname instance. We can't use it
the same way as a String when using it in `.start_with?` when comparing
backtrace lines for error causes in
`Transaction#first_formatted_backtrace_line`.

Fix this by casting the root_path always to a String in the Config
initializer so we're always sure it's a String and can use it as such.