Fix session data for Action Cable callbacks
In issue #1313 it was reported that session data was not reported for
Action Cable callbacks. This broke after PR #1209 where we changed the
way we handle sample data by allowing merging, which required more
strict type checks.

By calling `to_h`, ensure the value is a Hash, which is a valid sample
data type.

Fixes #1313