Send error causes line number as integer
The Ruby integration was sending this as a string, while the processor
expected a number. The processor has since been amended to accept
either a number or a string, but the Ruby integration should have
been sending a number to begin with.

See [Slack conversation][slack] and [processor PR][pr] for context.

[slack]: https://appsignal.slack.com/archives/C02A6H085/p1728551389181029
[pr]: https://github.com/appsignal/appsignal-processor-rs/pull/1577/files