Swap keys and values for option definitions
Like I mentioned in the previous commit, these constants aren't only
used to define how we read them from ENV vars. The ENV var name is less
important, so make it the value of these hashes.

If we want to add more config in the future, it's easier to structure
the values something like so:

`:option => { :env_key => "APPSIGNAL_<option name>", :required => true}`