Fix duplicate Railtie default config in specs
In the diagnose CLI spec the Railtie config defaults would be set twice,
failing the specs.

This was because the spec only checked beforehand if the MyApp constant
was present or not, which it is most of the time in this spec. Now also
check if the app is already initialized, so the Railtie isn't triggered
again, and then trigger it manually.