Move config load logic to its own method
Add a private method (that has to be publicly accessible within the gem
itself), to load the config. This avoids duplication in the diagnose CLI
when I update it to call this method as well.

Make it more safe by replicating the existing behavior: it will never
return a `nil` object for the Config once started. If it started with an
error or the config file is empty, initialize the config but set
`active` to `false`.