Add Ruby config file option to installer
To replace the YAML config file people need to know they can configure
the gem with a Ruby config file too. Add this option to the installer so
people can choose this config option when installing AppSignal in their
application.

I've made it the first listed method as it's the most flexible
configuration method. The YAML file option has the "legacy" label next
to it now to discourage people from using it.