Disable Hanami monkeypatch on Hanami 2.2
In the previous commit we started reading the action name from the
Hanami Action instance on the request env.

Disable the monkeypatch on Hanami 2.2+ so it's not doing the same thing
twice. We'll leave the monkeypatch for now, but eventually we should be
able to remove it. Like on the next major Hanami version or after some
years.

Part of #911