Add Appsignal.configure .env? helper
When adding environment specific config with an if-statement comparing
the current environment, they need to use a String. If apps compare a
String with a Symbol like `config.env == :production` it would return
false.

Apps can call `Appsignal.configure(:production)` with a Symbol and we'll
cast it to a String. The environment in the Config class is always a
String.

Add the `config.env?` helper to handle the casting of the environment
value so that our users don't have to bother with casting the
environment value to a String.