Update Hash representation in Strings
In Ruby 3.4 the internal representation of Hashes has changed when
called with `to_s`/`inspect`. Instead of `{:abc => :def}` it is now
`{abc: :def}`.

Update the Data struct spec to match both formats. We don't really rely
on the String representation of Data structs so it's fine.

For the ElasticSearch event formatter it does impact how we detect
unique events. It will no longer match the event between Ruby 3.3 and
Ruby 3.4 as the same event, meaning event metrics will be restarted
after upgrading to Ruby 3.4. This is already what happens upgrading to
Ruby 3.4 with published versions of the Ruby gem. And adding our own
formatter for arbitrary search parameters seems brittle. We'd have to
keep the Ruby format prior to Ruby 3.4 too.