Add ActiveSupport tagged logging support
Our logger did not support Rails ActiveSupport tagged logging yet.
This was because our logger didn't listen to the `tagged`, `push_tags`
and `pop_tags` methods.

Implement these methods to have the logger that wraps our logger call
these methods to set tags.

The format is the same as Rails formats it: `[My tag value]`
I decide not to set these as attributes, because they're only values.
They have no key and the value can be anything and change frequently.